refactor: entity 디렉토리 사용하는 것으로 통일
