+++ backend/src/map/entity/map-place.entity.ts
@@ -1,6 +1,6 @@
 import { Entity, Column, ManyToOne, JoinColumn } from 'typeorm';
-import { BaseEntity } from '../common/BaseEntity';
-import { Place } from '../place/place.entity';
+import { BaseEntity } from '../../common/BaseEntity';
+import { Place } from '../../place/entity/place.entity';
 import { Map } from './map.entity';
 
 @Entity()
