+++ backend/src/user/user.module.ts
@@ -3,7 +3,6 @@ import { UserController } from './user.controller';
 import { UserService } from './user.service';
 import { UserRepository } from './user.repository';
 
-// Todo. 유저 기능 완성 후 교체
 @Module({
   controllers: [UserController],
   providers: [UserService, UserRepository],
